requirements-dev.txt
factory-boy>=2.8.1,<2.9
